Move seq2seq output transform to it's own file
