+++ tests/text/seq2seq/translation/test_model.py
@@ -23,7 +23,7 @@ from flash.core.integrations.transformers.input_transform import TransformersInp
 from flash.core.utilities.imports import _TEXT_AVAILABLE
 from flash.text import TranslationTask
 from flash.text.input import TextDeserializer
-from flash.text.seq2seq.core.data import Seq2SeqOutputTransform
+from flash.text.seq2seq.core.output_transform import Seq2SeqOutputTransform
 from tests.helpers.utils import _SERVE_TESTING, _TEXT_TESTING
 
 # ======== Mock functions ========
